Refactor setup-environment.sh for clarity and functionality
The commit removes the on-load message from the `main` function and wraps it into a more distinct `about_script` function for improved readability. It also introduces a `must_be_sourced_msg` function to guide users on correct script execution. Lastly, it uses conditionals to ensure that the script is sourced, not executed.